Implement inverse dynamics in task space control
The changes consist of implementing an optional inverse dynamics control strategy in the task space control loop. This option is controlled by a preprocessor directive "ENABLE_INVERSE_DYNAMICS". If set to 1, the code computes the inverse dynamics of the system to produce control torques. If set to 0, it reverts to the earlier strategy. GainMatrices constructor has also been modified to allow varied gains. Other minor edits include relocating and updating function calls.